Refactor i3extractor.py into separate modules
