+++ src/graphnet/data/extractors/__init__.py
@@ -1,0 +1,4 @@
+from .i3extractor import *
+from .i3featureextractor import *
+from .i3truthextractor import *
+from .i3retroextractor import *
